requirements.txt
PyAudio==0.2.11
python-osc==1.6.3
pyzmq==16.0.2
